+++ test/test_gaussian.py
@@ -108,6 +108,71 @@ def test_empirical_bures_wasserstein_distance(nx, bias):
     np.testing.assert_allclose(10 * bias, nx.to_numpy(Wb), rtol=1e-2, atol=1e-2)
 
 
+def test_bures_wasserstein_barycenter(nx):
+    n = 50
+    k = 10
+    X = []
+    y = []
+    m = []
+    C = []
+    for _ in range(k):
+        X_, y_ = make_data_classif('3gauss', n)
+        m_ = np.mean(X_, axis=0)[None, :]
+        C_ = np.cov(X_.T)
+        X.append(X_)
+        y.append(y_)
+        m.append(m_)
+        C.append(C_)
+    m = np.array(m)
+    C = np.array(C)
+    X = nx.from_numpy(*X)
+    m = nx.from_numpy(m)
+    C = nx.from_numpy(C)
+
+    mblog, Cblog, log = ot.gaussian.bures_wasserstein_barycenter(m, C, log=True)
+    mb, Cb = ot.gaussian.bures_wasserstein_barycenter(m, C, log=False)
+
+    np.testing.assert_allclose(Cb, Cblog, rtol=1e-2, atol=1e-2)
+    np.testing.assert_allclose(mb, mblog, rtol=1e-2, atol=1e-2)
+
+    # Test weights argument
+    weights = nx.ones(k) / k
+    mbw, Cbw = ot.gaussian.bures_wasserstein_barycenter(m, C, weights=weights, log=False)
+    np.testing.assert_allclose(Cbw, Cb, rtol=1e-2, atol=1e-2)
+
+    # test with closed form for diagonal covariance matrices
+    Cdiag = [nx.diag(nx.diag(C[i])) for i in range(k)]
+    Cdiag = nx.stack(Cdiag, axis=0)
+    mbdiag, Cbdiag = ot.gaussian.bures_wasserstein_barycenter(m, Cdiag, log=False)
+
+    Cdiag_sqrt = [nx.sqrtm(C) for C in Cdiag]
+    Cdiag_sqrt = nx.stack(Cdiag_sqrt, axis=0)
+    Cdiag_mean = nx.mean(Cdiag_sqrt, axis=0)
+    Cdiag_cf = Cdiag_mean @ Cdiag_mean
+
+    np.testing.assert_allclose(Cbdiag, Cdiag_cf, rtol=1e-2, atol=1e-2)
+
+
+@pytest.mark.parametrize("bias", [True, False])
+def test_empirical_bures_wasserstein_barycenter(nx, bias):
+    n = 50
+    k = 10
+    X = []
+    y = []
+    for _ in range(k):
+        X_, y_ = make_data_classif('3gauss', n)
+        X.append(X_)
+        y.append(y_)
+
+    X = nx.from_numpy(*X)
+
+    mblog, Cblog, log = ot.gaussian.empirical_bures_wasserstein_barycenter(X, log=True, bias=bias)
+    mb, Cb = ot.gaussian.empirical_bures_wasserstein_barycenter(X, log=False, bias=bias)
+
+    np.testing.assert_allclose(Cb, Cblog, rtol=1e-2, atol=1e-2)
+    np.testing.assert_allclose(mb, mblog, rtol=1e-2, atol=1e-2)
+
+
 @pytest.mark.parametrize("d_target", [1, 2, 3, 10])
 def test_gaussian_gromov_wasserstein_distance(nx, d_target):
     ns = 400