+++ RELEASES.md
@@ -21,6 +21,8 @@
 + Add `stop_criterion` feature to (un)regularized (f)gw barycenter solvers (PR #578)
 + Add `fixed_structure` and `fixed_features` to entropic fgw barycenter solver (PR #578)
 + Add new entropic BAPG solvers for GW and FGW (PR #581)
++ Add Bures-Wasserstein barycenter in `ot.gaussian` (PR #582)
+
 
 #### Closed issues
 - Fix line search evaluating cost outside of the interpolation range (Issue #502, PR #504)